first pass at plugin command refactor
